[UXIT-1889] Scout PR · Move Governance Calendar Cards Logic [skip percy] (#920)
* Move CalelanderCards to local Governance folder

* Move Calender component to own file

* fixup! Move CalelanderCards to local Governance folder

* Abstract CalendarCard

* User EventType

* Move schemas

* Move getCalendarEvents
